update: Extract dependencies to doctrine from AnnotationScanner
